Fix clippy issues with strict settings
- Collapse nested if statements in decoder.rs (exponent/magnitude checks)
- Remove unused UnicodeNormalization import in lib.rs
- Derive Default on EncoderConfig instead of manual impl
- Add #[allow(dead_code)] to unused is_reserved() in type_dispatch_bench
- Prefix unused sign parameter with underscore in normalize_bignumber()
- Simplify char comparison to array pattern in conformance tests
- Collapse match into if let pattern for unicode_normalization option
- Remove unnecessary type cast in benches/comparison.rs
- Use struct initialization instead of field reassignment in quick_bench

+++ benches/comparison.rs
@@ -64,7 +64,7 @@ fn create_nested_data() -> Vec<ComplexStruct> {
             id: i as u64,
             name: format!("User {}", i),
             email: format!("user{}@example.com", i),
-            scores: vec![i as i32; 10],
+            scores: vec![i; 10],
             metadata: Metadata {
                 created: "2024-01-15T10:30:00Z".to_string(),
                 updated: "2024-01-18T14:22:33Z".to_string(),
